refactor: put web server in its own package
